bakery_grpc/requirements.txt
grpcio
grpcio-tools
pydantic
python-dotenv
grpc-interceptor